api/requirements.txt
APScheduler==3.10.1
blinker==1.5
click==8.1.3
Flask==2.2.3
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
PyJWT==2.6.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
six==1.16.0
tzdata==2022.7
tzlocal==4.2
Werkzeug==2.2.3
